Monetary variables from Double to BigDecimal data type

+++ server/api/src/main/java/com/twb/pokerapp/domain/PlayerSession.java
@@ -11,6 +11,7 @@ import lombok.Setter;
 import org.apache.commons.lang3.builder.EqualsBuilder;
 import org.apache.commons.lang3.builder.HashCodeBuilder;
 
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.UUID;
@@ -49,7 +50,7 @@ public class PlayerSession extends Auditable {
     private Boolean active;
 
     @Column(name = "funds")
-    private Double funds;
+    private BigDecimal funds;
 
     @NotNull
     @Enumerated(EnumType.STRING)
